@@ -36,47 +36,47 @@
     "*.ts": "eslint --cache --fix"
   },
   "dependencies": {
-    "@angular/animations": "~13.3.7",
-    "@angular/cdk": "~13.3.7",
-    "@angular/common": "~13.3.7",
-    "@angular/compiler": "~13.3.7",
-    "@angular/core": "~13.3.7",
-    "@angular/forms": "~13.3.7",
-    "@angular/material": "~13.3.7",
-    "@angular/platform-browser": "~13.3.7",
-    "@angular/platform-browser-dynamic": "~13.3.7",
-    "@angular/router": "~13.3.7",
-    "ang-jsoneditor": "^1.10.5",
-    "deep-equal": "^2.0.5",
+    "@angular/animations": "~15.2.0",
+    "@angular/cdk": "~15.2.0",
+    "@angular/common": "~15.2.0",
+    "@angular/compiler": "~15.2.0",
+    "@angular/core": "~15.2.0",
+    "@angular/forms": "~15.2.0",
+    "@angular/material": "~15.2.0",
+    "@angular/platform-browser": "~15.2.0",
+    "@angular/platform-browser-dynamic": "~15.2.0",
+    "@angular/router": "~15.2.0",
+    "ang-jsoneditor": "^2.0.0",
+    "deep-equal": "^2.2.0",
     "file-saver": "^2.0.5",
     "http-status-codes": "^2.2.0",
     "js-yaml": "^4.1.0",
-    "jsoneditor": "^9.7.4",
+    "jsoneditor": "^9.10.0",
     "jsonpath": "^1.1.1",
     "lodash.clonedeep": "^4.5.0",
-    "ngx-logger": "^5.0.9",
-    "openapi-schema-validator": "^11.0.0",
-    "openapi-types": "^11.0.0",
-    "prettier": "^2.6.2",
-    "rxjs": "~7.5.5",
+    "ngx-logger": "^5.0.11",
+    "openapi-schema-validator": "^12.1.0",
+    "openapi-types": "^12.1.0",
+    "prettier": "^2.8.4",
+    "rxjs": "~7.8.0",
     "rxjs-observable-store": "^3.1.0",
-    "tslib": "^2.4.0",
-    "uuid": "^8.3.2",
-    "zone.js": "~0.11.5"
+    "tslib": "^2.5.0",
+    "uuid": "^9.0.0",
+    "zone.js": "~0.12.0"
   },
   "devDependencies": {
-    "@angular-builders/jest": "^13.0.4",
-    "@angular-devkit/architect": "^0.1303.5",
-    "@angular-devkit/build-angular": "~13.3.5",
+    "@angular-builders/jest": "^15.0.0",
+    "@angular-devkit/architect": "^0.1502.0",
+    "@angular-devkit/build-angular": "~15.2.0",
     "@angular-eslint/builder": "15.2.1",
     "@angular-eslint/eslint-plugin": "15.2.1",
     "@angular-eslint/eslint-plugin-template": "15.2.1",
-    "@angular-eslint/schematics": "13.2.1",
+    "@angular-eslint/schematics": "15.2.1",
     "@angular-eslint/template-parser": "15.2.1",
-    "@angular/cli": "~13.3.5",
-    "@angular/compiler-cli": "~13.3.7",
-    "@angular/elements": "^13.3.7",
-    "@angular/language-service": "~13.3.7",
+    "@angular/cli": "~15.2.0",
+    "@angular/compiler-cli": "~15.2.0",
+    "@angular/elements": "^15.2.0",
+    "@angular/language-service": "~15.2.0",
     "@babel/core": "^7.21.0",
     "@compodoc/compodoc": "^1.1.19",
     "@faker-js/faker": "^7.6.0",
@@ -90,13 +90,13 @@
     "@types/file-saver": "^2.0.5",
     "@types/jasmine": "^4.3.1",
     "@types/jasminewd2": "~2.0.10",
-    "@types/jest": "^27.5.1",
+    "@types/jest": "^29.4.0",
     "@types/js-yaml": "^4.0.5",
     "@types/lodash.clonedeep": "^4.5.7",
-    "@types/node": "^18.14.0",
-    "@types/uuid": "^9.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.23.0",
-    "@typescript-eslint/parser": "^5.23.0",
+    "@types/node": "^18.14.1",
+    "@types/uuid": "^9.0.1",
+    "@typescript-eslint/eslint-plugin": "^5.53.0",
+    "@typescript-eslint/parser": "^5.53.0",
     "@webcomponents/custom-elements": "^1.5.1",
     "babel-loader": "^9.1.2",
     "eslint": "^8.34.0",
@@ -108,15 +108,15 @@
     "jasmine-core": "^4.5.0",
     "jasmine-reporters": "^2.5.2",
     "jasmine-spec-reporter": "~7.0.0",
-    "jest": "^27.5.1",
+    "jest": "^29.4.3",
     "jest-dashboard": "^0.3.1",
-    "jest-junit": "^13.2.0",
-    "jest-preset-angular": "^11.1.2",
+    "jest-junit": "^15.0.0",
+    "jest-preset-angular": "^13.0.0",
     "lint-staged": "^13.1.2",
     "package-lock-check": "^1.1.0",
     "protractor": "~7.0.0",
-    "ts-jest": "^27.1.4",
+    "ts-jest": "^29.0.5",
     "ts-node": "^10.9.1",
-    "typescript": "~4.6.4"
+    "typescript": "~4.9.5"
   }
 }
